Добавлено уведомление о нецспешном входе, добавлена новая таблица Payments, начато приобразование адаптера для вывода платежей с бд

+++ app/src/main/java/com/example/kursovaya/ItemHistory.kt
@@ -1,3 +1,9 @@
 package com.example.kursovaya
 
-data class ItemHistory(val idPayment: Int, val indexPay: Boolean, val sum: Int, val title: String, val date: String)
+data class ItemHistory(
+    val idPayment: Int,
+    val indexPay: Boolean,
+    val sum: Double,
+    val title: String,
+    val date: String
+)
